starting point for bdvhandle filter nodes
